refactor: move media settings to developer tab, centralize config defaults
